Save book to JSON file

+++ book_logic.rb
@@ -11,21 +11,39 @@ class BookLogic
     @books = Data.books
   end
 
-  def add_to_data
+  def add_to_data(title, author)
+    path = './store/books.json'
+    file = File.new(path, 'w')
+    File.write(path, '[]')
+    file = File.open(path, 'r')
+
+    # loaded_books = JSON.parse(file.read)
+
+    loaded_books = []
+
+    @books.each do |book|
+      loaded_books << {"title" => book.title, "author" => book.author}
+    end
+
+
+    loaded_books << {"title" => title, "author" => author}
+
+    File.open(path, 'w') do |f|
+      f.write(loaded_books.to_json) 
+    end      
+
   end
 
   def load_book_data
     path = './store/books.json'
-    if File.exists?(path)
-      file = File.open(path , 'r')
-    loaded_books = JSON.parse(file.read)
-    loaded_books.each do |book|
-      book = Book.new(book['title'], book['author'])
-      @books.push(book)
-    end
-  else 
-    file = File.new(path, 'w')
-    file.write(@books.to_json)
+    if File.exists?(path) && File.size(path) != 2
+          file = File.open(path , 'r')
+          loaded_books = JSON.parse(file.read)
+          loaded_books.each do |b|
+          book = Book.new(b['title'], b['author'])
+          @books.push(book)
+          end
+          # @books = []
     end 
   end
 
@@ -48,6 +66,7 @@ class BookLogic
     print 'Author: '
     author = gets.chomp.capitalize
     book = Book.new(title, author)
+    add_to_data(title, author)
     @books.push(book)
     puts 'Book created successfully'
   end
